perf(mc): code-split the G-1114.D Network graph engine into a lazy chunk
PR #905 review FIX-FIRST: the dashboard's default view is the working grid,
not Network, so every principal was downloading + parsing xyflow + the ELK
engine (+0.62 MB gzip) in the entry bundle for a tab they may never open.

Split the heavy half into a lazily-loaded chunk:
- New components/network-canvas.tsx — the ONLY module importing @xyflow/react +
  elkjs. Owns the ELK layout effect + the React Flow canvas. Default export so
  React.lazy resolves it.
- network-view.tsx drops the xyflow/elk imports; keeps the chrome
  (heading/subtitle/empty/loading/error) + the pure buildNetworkGraph adapter
  in the main bundle, and React.lazy-imports the canvas behind <Suspense>
  (fallback "Loading network…") gated on mode==="list".
- package.json build:dashboard + watch:dashboard gain --splitting. This is
  load-bearing: bun does NOT code-split dynamic import() by default for the
  HTML/browser target — without the flag it inlines the canvas back into the
  single entry bundle. Docs (dashboard-deployment, design-mission-control)
  updated to match.

Verified split via `bun run build:dashboard`:
- entry bundle: 4.78 MB raw / 0.84 MB gz  →  1.21 MB raw / 0.21 MB gz
  (+ a 71 KB shared chunk) — back to the pre-graph baseline; the common path
  no longer pays for the engine.
- new network-canvas-*.js chunk: 3.50 MB raw / 0.62 MB gz (xyflow + ELK),
  + network-canvas-*.css 18 KB — both load only on first Network-tab entry.

Tests: network-view smoke now asserts the Suspense fallback ("Loading
network…") under static markup (the lazy chunk never resolves synchronously);
the adapter/layout/node tests are unchanged (they import directly). Full suite
5623 pass / 0 fail; tsc clean; lint 0 errors; carveouts PASS.

Deferred nits: per-heartbeat ELK re-layout filed on the #355 umbrella
(harmless today — debounce + sub-ms radial + latest-wins guard); the
untested-wrapper split was assessed clean by the reviewer (no action).

refs #355
