+++ package.json
@@ -15,8 +15,8 @@
     }
   },
   "scripts": {
-    "build:dashboard": "bun build src/surface/mc/dashboard-v2/index.html --outdir dist/dashboard-v2 --target browser",
-    "watch:dashboard": "bun build src/surface/mc/dashboard-v2/index.html --outdir dist/dashboard-v2 --target browser --watch",
+    "build:dashboard": "bun build src/surface/mc/dashboard-v2/index.html --outdir dist/dashboard-v2 --target browser --splitting",
+    "watch:dashboard": "bun build src/surface/mc/dashboard-v2/index.html --outdir dist/dashboard-v2 --target browser --splitting --watch",
     "lint": "eslint .",
     "lint:errors": "eslint --quiet .",
     "lint:fix": "eslint --fix ."
